refactor: fee support on fixtures

+++ src/interfaces/IBundleRunner.sol
@@ -18,7 +18,7 @@ pragma solidity ^0.8.22;
 interface IBundleRunner {
     struct BundleExecutionParams {
         address bundle;
-        uint transactionCost;
+        uint256 transactionCost;
     }
 
     function runBundles(BundleExecutionParams[] calldata _bundles) external;
